pyproject.toml
[tool.poetry]
name = "relishplus"
version = "0.1.0"
description = "Automated meal ordering system for Relish/ezCater"
authors = ["Gary Sheng <gary.sheng@gauntletai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
browser-use = "^0.1.0"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
typer = "^0.9.0"
rich = "^13.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 